Enemy Patrol Script
Enemy Agents will move from one transform position to another transform position.
The array of Transforms is expected to be empty GameObjects that will be "Nodes" scattered around the scene.
the Agent is currently moving by using unity's in built NavMesh System. A navMesh must be created in the scene and the agent GameObject requires NavMeshAgent component