Example Movement Prototype Scene
This scene contains one patrolling agent and one chasing Agent.
The patroller moves from one node to another. It waits on arrival. then turns and then continues.
The chaser moves around obstacles towards the patroller as a target.
This scene MUST contain a NavMesh in order for the agents to move around.