Add PlayerControl.cs
This script contains movement for the player.
It moves position depending on the referenced transform.
Currently there are 3 different transforms:
CameraAxis (The parent transform the camera is attached to)
Global (No transform is referenced)
Player (Player transform references it's self)
Currently if the player transform is referencing it's self, it will simulate tank controls. However, it is optimised poorly and does not work correctly, therefore it is only a small example of how that might work.