Example of Vision behind wall
Updated the scene to include the Agent tag.
A ray is sent between the vision object and the moving object. If there is any other object with the Agent tag that is not the player and the ray has collided with them. then they have not been seen. Currently this is working backwards, as the ray origin is the mover at the moment. it can still serve as an example.